+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [`x.y.z`] - Unreleased
 
+## [`9.111.8`] - 2020-06-15
+
 ### New Known Issues:
 
 ### Breaking Changes: